requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
langchain-core==0.3.55
langchain-openai==0.3.14
python-dotenv